@ReneWerner87 2025-01-19
@miyamo2 thanks for the customization

can you please expand the documentation of the caching resources and mention that we are now following RFC7231 and link to it so that more info can be found ?
and in the what's new markdown it would be good if we mention that the middleware has now been improved


- [ ] update cache mw markdown
- [ ] update what´s new markdown